first training

+++ gender_recognition/train.py
@@ -416,7 +416,7 @@ def dataio_prep(hparams):
         This is done on the CPU in the `collate_fn`.
         It resamples the signal if the sampling rate is not the same as in the hyperparameter file.
         """
-        sig, read_sr = torchaudio.load(wav, frame_offset=16000*start, num_frames=16000*end)
+        sig, read_sr = torchaudio.load(wav, frame_offset=int(16000*start), num_frames=int(16000*end))
 
         # If multi-channels, downmix it to a mono channel
         sig = torch.squeeze(sig)
